+++ Classes/Model/Type/Atlas.php
@@ -87,7 +87,6 @@ final class Atlas extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',