+++ Classes/Model/Type/Chapter.php
@@ -90,7 +90,6 @@ final class Chapter extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',