+++ Classes/Model/Type/ComicIssue.php
@@ -15,10 +15,6 @@ use Brotkrueml\Schema\Core\Model\AbstractType;
 
 /**
  * Individual comic issues are serially published as
- * part of a larger series. For the sake of consistency, even one-shot issues
- * belong to a series comprised of a single issue. All comic issues can be
- * uniquely identified by: the combination of the name and volume number of the
- * series to which the issue belongs; the issue number; and the variant
  * description of the issue (if any).
  */
 final class ComicIssue extends AbstractType
@@ -101,7 +97,6 @@ final class ComicIssue extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',