+++ Classes/Model/Type/ComicStory.php
@@ -15,7 +15,6 @@ use Brotkrueml\Schema\Core\Model\AbstractType;
 
 /**
  * The term "story" is any indivisible, re-printable
- * unit of a comic, including the interior stories, covers, and backmatter. Most
  * comics have at least two stories: a cover (ComicCoverArt) and an interior story.
  */
 final class ComicStory extends AbstractType
@@ -94,7 +93,6 @@ final class ComicStory extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',