+++ Classes/Model/Type/CoverArt.php
@@ -98,7 +98,6 @@ final class CoverArt extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',