+++ Classes/Model/Type/Newspaper.php
@@ -89,7 +89,6 @@ final class Newspaper extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',