+++ Classes/Model/Type/Thesis.php
@@ -88,7 +88,6 @@ final class Thesis extends AbstractType
         'position',
         'potentialAction',
         'producer',
-        'provider',
         'publication',
         'publisher',
         'publisherImprint',