+++ Classes/ViewHelpers/Type/CollectionViewHelper.php
@@ -14,7 +14,7 @@ namespace Brotkrueml\SchemaBib\ViewHelpers\Type;
 use Brotkrueml\Schema\Core\ViewHelpers\AbstractTypeViewHelper;
 
 /**
- * A collection of items e.g. creative works or products.
+ * A collection of items, e.g. creative works or products.
  */
 final class CollectionViewHelper extends AbstractTypeViewHelper
 {