+++ Classes/ViewHelpers/Type/ComicStoryViewHelper.php
@@ -15,7 +15,6 @@ use Brotkrueml\Schema\Core\ViewHelpers\AbstractTypeViewHelper;
 
 /**
  * The term "story" is any indivisible, re-printable
- * unit of a comic, including the interior stories, covers, and backmatter. Most
  * comics have at least two stories: a cover (ComicCoverArt) and an interior story.
  */
 final class ComicStoryViewHelper extends AbstractTypeViewHelper